feat: avoid head request unless necessary (#6)
This avoids an additional HEAD request unless it absolutely needed.

It's needed when you request a open range (`100-` for example).

The potentially breaking change (except it's not since we don't rely on it) is that for multipart requests, the `Content-Range` response header now does not include the total length unless you've asked for an open-ended range. i.e. you used to get `Content-Range: bytes 1-3/8` and you now get `Content-Range: bytes 1-3/*`.